revert: クロスポスト機能 (#626) を取り下げ
This reverts commit 61fb7b2b6dbde1053ec2eb63e1fddbd93ba86e51.

カスタム絵文字の命名規則がサーバーごとに異なる、投稿の成功可否が
複数サーバーで別々に判定されるなど、サーバー固有の事情で支障が出る
パターンが多すぎて実運用に耐えないと判断した。

クロスアカウント基盤 (#777) 自体と「別のアカウントで…」メニュー
(#627) は維持する — 取り下げるのは複数アカウント同時投稿のみ。
